+++ fibber/datasets/dataset_utils.py
@@ -1,6 +1,6 @@
 """This module provides utility functions and classes to handle fibber's datasets.
 
-* To load a dataset, use `get_dataset` function. For example, to load AG's news dataset, run::
+* To load a dataset, use ``get_dataset`` function. For example, to load AG's news dataset, run::
 
     trainset, testset =  get_dataset("ag")
 
@@ -47,9 +47,9 @@ import os
 import numpy as np
 import torch
 import tqdm
-from transformers import BertTokenizer
+from transformers import BertTokenizerFast
 
-from fibber import get_root_dir, log
+from fibber import get_root_dir, log, resources
 from fibber.datasets.downloadable_datasets import downloadable_dataset_urls
 from fibber.download_utils import download_file
 
@@ -86,7 +86,7 @@ def get_dataset(dataset_name):
 
     if not os.path.exists(train_filename) or not os.path.exists(test_filename):
         logger.error("%s dataset not found.", dataset_name)
-        assert 0, ("Please use `python3 -m fibber.benchmark.download_datasets` "
+        assert 0, ("Please use `python3 -m fibber.datasets.download_datasets` "
                    "to download datasets.")
 
     with open(train_filename) as f:
@@ -259,7 +259,8 @@ class DatasetForBert(torch.utils.data.IterableDataset):
         self._data = [[] for i in range(len(self._buckets))]
 
         self._batch_size = batch_size
-        self._tokenizer = BertTokenizer.from_pretrained(model_init)
+        self._tokenizer = BertTokenizerFast.from_pretrained(
+            resources.get_transformers(model_init), do_lower_case="uncased" in model_init)
 
         self._seed = seed
         self._pad_tok_id = self._tokenizer.pad_token_id
@@ -335,7 +336,7 @@ class DatasetForBert(torch.utils.data.IterableDataset):
                 masked_pos[:, 0] = 0
                 lm_labels = (masked_pos * texts - 100 * (1 - masked_pos))
                 rand_t = self._rng.rand(self._batch_size, max_text_len)
-                filling = self._rng.randint(0, len(self._tokenizer),
+                filling = self._rng.randint(0, self._tokenizer.vocab_size,
                                             (self._batch_size, max_text_len))
                 filling = ((rand_t < 0.8) * self._mask_tok_id
                            + (rand_t >= 0.8) * (rand_t < 0.9) * filling