+++ fibber/resources/__init__.py
@@ -1,3 +1,13 @@
-from fibber.resources.resource_utils import get_glove_emb, get_stopwords
+from fibber.resources.download_resources import download_all
+from fibber.resources.resource_utils import (
+    get_corenlp, get_glove_emb, get_nltk_data, get_stopwords, get_transformers,
+    get_universal_sentence_encoder)
 
-__all__ = ["get_glove_emb", "get_stopwords"]
+__all__ = [
+    "get_glove_emb",
+    "get_stopwords",
+    "get_nltk_data",
+    "get_universal_sentence_encoder",
+    "get_transformers",
+    "get_corenlp",
+    "download_all"]