+++ HISTORY.md
@@ -1,5 +1,14 @@
 # History
 
+## version 0.1.0
+
+This release is a major update to Fibber library. Advanced paraphrase algorithms are included.
+
+- Add two strategies: TextFoolerStrategy and BertSamplingStrategy.
+- Improve the benchmarking framework: add more metrics specifically designed for adversarial attack.
+- Datasets: add a variation of AG's news dataset, `ag_no_title`.
+- Bug fix and improvements.
+
 ## version 0.0.1
 
 This is the first release of Fibber library. This release contains: